Module google_cloudprivatecatalogproducer1_beta1::api

Structs

CatalogAssociationCreateCall
Creates an Association instance under a given Catalog.

CatalogAssociationDeleteCall
Deletes the given Association.

CatalogAssociationGetCall
Returns the requested Association resource.

CatalogAssociationListCall
Lists all Association resources under a catalog.

CatalogCreateCall
Creates a new Catalog resource.

CatalogDeleteCall
Soft deletes an existing Catalog and all resources under it. The catalog can only be deleted if there is no associations under it or DeleteCatalogRequest.force is true. The delete operation can be recovered by the PrivateCatalogProducer.UndeleteCatalog method.

CatalogGetCall
Returns the requested Catalog resource.

CatalogGetIamPolicyCall
Gets IAM policy for the specified Catalog.

CatalogListCall
Lists Catalog resources that the producer has access to, within the scope of the parent resource.

CatalogMethods
A builder providing access to all methods supported on catalog resources. It is not used directly, but through the CloudPrivateCatalogProducer hub.

CatalogPatchCall
Updates a specific Catalog resource.

CatalogProductCopyCall
Copies a Product under another Catalog.

CatalogProductCreateCall
Creates a Product instance under a given Catalog.

CatalogProductDeleteCall
Hard deletes a Product.

CatalogProductGetCall
Returns the requested Product resource.

CatalogProductIconUploadCall
Creates an Icon instance under a given Product. If Product only has a default icon, a new Icon instance is created and associated with the given Product. If Product already has a non-default icon, the action creates a new Icon instance, associates the newly created Icon with the given Product and deletes the old icon.

CatalogProductListCall
Lists Product resources that the producer has access to, within the scope of the parent catalog.

CatalogProductPatchCall
Updates a specific Product resource.

CatalogProductVersionCreateCall
Creates a Version instance under a given Product.

CatalogProductVersionDeleteCall
Hard deletes a Version.

CatalogProductVersionGetCall
Returns the requested Version resource.

CatalogProductVersionListCall
Lists Version resources that the producer has access to, within the scope of the parent Product.

CatalogProductVersionPatchCall
Updates a specific Version resource.

CatalogSetIamPolicyCall
Sets the IAM policy for the specified Catalog.

CatalogTestIamPermissionCall
Tests the IAM permissions for the specified Catalog.

CatalogUndeleteCall
Undeletes a deleted Catalog and all resources under it.

CloudPrivateCatalogProducer
Central instance to access all CloudPrivateCatalogProducer related resource activities

GoogleCloudPrivatecatalogproducerV1beta1Association
An association tuple that pairs a Catalog to a resource that can use the Catalog. After association, a google.cloud.privatecatalog.v1beta1.Catalog becomes available to consumers under specified Association.resource and all of its child nodes. Users who have the cloudprivatecatalog.targets.get permission on any of the resource nodes can access the catalog and child products under the node.

GoogleCloudPrivatecatalogproducerV1beta1Catalog
The producer representation of a catalog which is a curated collection of solutions that can be managed, controlled, and shared by cloud admins.

GoogleCloudPrivatecatalogproducerV1beta1CopyProductRequest
There is no detailed description.

GoogleCloudPrivatecatalogproducerV1beta1CreateAssociationRequest
There is no detailed description.

GoogleCloudPrivatecatalogproducerV1beta1ListAssociationsResponse
There is no detailed description.

GoogleCloudPrivatecatalogproducerV1beta1ListCatalogsResponse
There is no detailed description.

GoogleCloudPrivatecatalogproducerV1beta1ListProductsResponse
There is no detailed description.

GoogleCloudPrivatecatalogproducerV1beta1ListVersionsResponse
There is no detailed description.

GoogleCloudPrivatecatalogproducerV1beta1Product
The producer representation of a product which is a child resource of Catalog with display metadata and a list of Version resources.

GoogleCloudPrivatecatalogproducerV1beta1UndeleteCatalogRequest
There is no detailed description.

GoogleCloudPrivatecatalogproducerV1beta1UploadIconRequest
There is no detailed description.

GoogleCloudPrivatecatalogproducerV1beta1Version
The producer representation of a version, which is a child resource under a Product with asset data.

GoogleIamV1AuditConfig
Specifies the audit configuration for a service. The configuration determines which permission types are logged, and what identities, if any, are exempted from logging. An AuditConfig must have one or more AuditLogConfigs.

GoogleIamV1AuditLogConfig
Provides the configuration for logging a type of permissions. Example:

GoogleIamV1Binding
Associates members with a role.

GoogleIamV1Policy
An Identity and Access Management (IAM) policy, which specifies access controls for Google Cloud resources.

GoogleIamV1SetIamPolicyRequest
Request message for SetIamPolicy method.

GoogleIamV1TestIamPermissionsRequest
Request message for TestIamPermissions method.

GoogleIamV1TestIamPermissionsResponse
Response message for TestIamPermissions method.

GoogleLongrunningCancelOperationRequest
The request message for Operations.CancelOperation.

GoogleLongrunningListOperationsResponse
The response message for Operations.ListOperations.

GoogleLongrunningOperation
This resource represents a long-running operation that is the result of a network API call.

GoogleProtobufEmpty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance:

GoogleRpcStatus
The Status type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by gRPC. Each Status message contains three pieces of data: error code, error message, and error details.

GoogleTypeExpr
Represents a textual expression in the Common Expression Language (CEL) syntax. CEL is a C-like expression language. The syntax and semantics of CEL are documented at https://github.com/google/cel-spec.

OperationCancelCall
Starts asynchronous cancellation on a long-running operation. The server makes a best effort to cancel the operation, but success is not guaranteed. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED. Clients can use Operations.GetOperation or other methods to check whether the cancellation succeeded or whether the operation completed despite cancellation. On successful cancellation, the operation is not deleted; instead, it becomes an operation with an Operation.error value with a google.rpc.Status.code of 1, corresponding to Code.CANCELLED.

OperationDeleteCall
Deletes a long-running operation. This method indicates that the client is no longer interested in the operation result. It does not cancel the operation. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED.

OperationGetCall
Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

OperationListCall
Lists operations that match the specified filter in the request. If the server doesn't support this method, it returns UNIMPLEMENTED.

OperationMethods
A builder providing access to all methods supported on operation resources. It is not used directly, but through the CloudPrivateCatalogProducer hub.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.